+++ modules/apps/foundation/portal-security/portal-security-membership-policy-test/src/testIntegration/java/com/liferay/portal/security/membership/policy/role/test/RoleMembershipPolicyBasicTest.java
@@ -1,0 +1,84 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.membership.policy.role.test;
+
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.portal.kernel.security.membershippolicy.RoleMembershipPolicyUtil;
+import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.security.membership.policy.role.BaseRoleMembershipPolicyTestCase;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.test.rule.PermissionCheckerTestRule;
+
+import org.junit.Assert;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Roberto Díaz
+ */
+@RunWith(Arquillian.class)
+public class RoleMembershipPolicyBasicTest
+	extends BaseRoleMembershipPolicyTestCase {
+
+	@ClassRule
+	@Rule
+	public static final AggregateTestRule aggregateTestRule =
+		new AggregateTestRule(
+			new LiferayIntegrationTestRule(),
+			PermissionCheckerTestRule.INSTANCE);
+
+	@Test
+	public void testIsRoleAllowed() throws Exception {
+		long[] userIds = addUsers();
+		long[] standardRoleIds = addStandardRoles();
+
+		Assert.assertTrue(
+			RoleMembershipPolicyUtil.isRoleAllowed(
+				userIds[0], standardRoleIds[0]));
+	}
+
+	@Test
+	public void testIsRoleNotAllowed() throws Exception {
+		long[] userIds = addUsers();
+		long[] forbiddenRoleIds = addForbiddenRoles();
+
+		Assert.assertFalse(
+			RoleMembershipPolicyUtil.isRoleAllowed(
+				userIds[0], forbiddenRoleIds[0]));
+	}
+
+	@Test
+	public void testIsRoleNotRequired() throws Exception {
+		long[] userIds = addUsers();
+		long[] standardRoleIds = addStandardRoles();
+
+		Assert.assertFalse(
+			RoleMembershipPolicyUtil.isRoleRequired(
+				userIds[0], standardRoleIds[0]));
+	}
+
+	@Test
+	public void testIsRoleRequired() throws Exception {
+		long[] userIds = addUsers();
+		long[] requiredRoleIds = addRequiredRoles();
+
+		Assert.assertTrue(
+			RoleMembershipPolicyUtil.isRoleRequired(
+				userIds[0], requiredRoleIds[0]));
+	}
+
+}