+++ modules/util/css-builder/src/test/java/com/liferay/css/builder/CSSBuilderTest.java
@@ -14,35 +14,31 @@
 
 package com.liferay.css.builder;
 
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.css.builder.util.FileTestUtil;
 
 import java.io.File;
-import java.io.IOException;
 
 import java.net.URL;
 
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.junit.After;
 import org.junit.Assert;
+import org.junit.Before;
 import org.junit.BeforeClass;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
 
 /**
  * @author Eduardo Garcia
  * @author David Truong
+ * @author Andrea Di Giorgi
  */
 public class CSSBuilderTest {
 
@@ -50,107 +46,98 @@ public class CSSBuilderTest {
 	public static void setUpClass() throws Exception {
 		URL url = CSSBuilderTest.class.getResource("dependencies");
 
-		Path path = Paths.get(url.toURI());
+		_dependenciesDirPath = Paths.get(url.toURI());
 
-		_docrootDirName = path.toString();
+		Assert.assertTrue(Files.isDirectory(_dependenciesDirPath));
+
+		_portalCommonDirPath = Paths.get("build/portal-common-css");
+
+		Assert.assertTrue(Files.isDirectory(_portalCommonDirPath));
+
+		_portalCommonJarPath = Paths.get(
+			"build/portal-common-css-jar/com.liferay.frontend.css.common.jar");
+
+		Assert.assertTrue(Files.isRegularFile(_portalCommonJarPath));
 	}
 
-	@After
-	public void tearDown() throws Exception {
-		Files.walkFileTree(
-			Paths.get(_docrootDirName + "/css/.sass-cache"),
-			new SimpleFileVisitor<Path>() {
+	@Before
+	public void setUp() throws Exception {
+		File docrootDir = temporaryFolder.getRoot();
 
-				@Override
-				public FileVisitResult postVisitDirectory(
-						Path path, IOException ioe)
-					throws IOException {
+		_docrootDirPath = docrootDir.toPath();
 
-					Files.delete(path);
-
-					return FileVisitResult.CONTINUE;
-				}
-
-				@Override
-				public FileVisitResult visitFile(
-						Path path, BasicFileAttributes basicFileAttributes)
-					throws IOException {
-
-					Files.delete(path);
-
-					return FileVisitResult.CONTINUE;
-				}
-
-			});
+		FileTestUtil.copyDir(
+			_dependenciesDirPath.resolve("css"),
+			_docrootDirPath.resolve("css"));
 	}
 
 	@Test
-	public void testCssBuilderWithFragmentChange() throws Exception {
-		Path fragmentFileToChange = Paths.get(
-			_docrootDirName, "css", "_import_change.scss");
+	public void testCSSBuilderWithFragmentChange() throws Exception {
+		Path fragmentChangePath = _docrootDirPath.resolve(
+			"css/_import_change.scss");
 
-		_changeContentInPath(fragmentFileToChange, "brown", "khaki");
+		FileTestUtil.changeContentInPath(fragmentChangePath, "brown", "khaki");
 
-		try (CSSBuilder cssBuilder = new CSSBuilder(
-				_docrootDirName, false, ".sass-cache/",
-				_PORTAL_COMMON_CSS_DIR_NAME, 6, new String[0], "jni")) {
+		executeCSSBuilder(
+			"/css", _docrootDirPath, false, ".sass-cache/",
+			_portalCommonDirPath, 6, new String[0], "jni");
 
-			cssBuilder.execute(Arrays.asList(new String[] {"/css"}));
-		}
+		Path cssPath = _docrootDirPath.resolve(
+			"css/.sass-cache/test_import_change.css");
 
-		String outputCssFilePath =
-			_docrootDirName + "/css/.sass-cache/test_import_change.css";
+		String css = FileTestUtil.read(cssPath);
 
-		String outputCssFileContent = _read(outputCssFilePath);
+		FileTestUtil.changeContentInPath(fragmentChangePath, "khaki", "brown");
 
-		_changeContentInPath(fragmentFileToChange, "khaki", "brown");
+		executeCSSBuilder(
+			"/css", _docrootDirPath, false, ".sass-cache/",
+			_portalCommonDirPath, 6, new String[0], "jni");
 
-		try (CSSBuilder cssBuilder = new CSSBuilder(
-				_docrootDirName, false, ".sass-cache/",
-				_PORTAL_COMMON_CSS_DIR_NAME, 6, new String[0], "jni")) {
+		css = FileTestUtil.read(cssPath);
 
-			cssBuilder.execute(Arrays.asList(new String[] {"/css"}));
-		}
-
-		outputCssFileContent = _read(outputCssFilePath);
-
-		Assert.assertTrue(outputCssFileContent.contains("brown"));
+		Assert.assertTrue(css.contains("brown"));
 	}
 
 	@Test
-	public void testCssBuilderWithJni() throws Exception {
-		_testCssBuilder("jni", _PORTAL_COMMON_CSS_DIR_NAME);
+	public void testCSSBuilderWithJni() throws Exception {
+		_testCSSBuilder(_portalCommonDirPath, "jni");
 	}
 
 	@Test
-	public void testCssBuilderWithJniAndPortalCommonJar() throws Exception {
-		_testCssBuilder("jni", _PORTAL_COMMON_CSS_JAR_FILE_NAME);
+	public void testCSSBuilderWithJniAndPortalCommonJar() throws Exception {
+		_testCSSBuilder(_portalCommonJarPath, "jni");
 	}
 
 	@Test
-	public void testCssBuilderWithRuby() throws Exception {
-		_testCssBuilder("ruby", _PORTAL_COMMON_CSS_DIR_NAME);
+	public void testCSSBuilderWithRuby() throws Exception {
+		_testCSSBuilder(_portalCommonDirPath, "ruby");
 	}
 
 	@Test
-	public void testCssBuilderWithRubyAndPortalCommonJar() throws Exception {
-		_testCssBuilder("ruby", _PORTAL_COMMON_CSS_DIR_NAME);
+	public void testCSSBuilderWithRubyAndPortalCommonJar() throws Exception {
+		_testCSSBuilder(_portalCommonJarPath, "ruby");
 	}
 
-	private static void _changeContentInPath(
-			Path path, String s, String replacement)
+	@Rule
+	public TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+	protected void executeCSSBuilder(
+			String dirName, Path docrootDirPath, boolean generateSourceMap,
+			String outputDirName, Path portalCommonPath, int precision,
+			String[] rtlExcludedPathRegexps, String sassCompilerClassName)
 		throws Exception {
 
-		Charset charset = StandardCharsets.UTF_8;
+		try (CSSBuilder cssBuilder = new CSSBuilder(
+				String.valueOf(docrootDirPath.toAbsolutePath()),
+				generateSourceMap, outputDirName,
+				String.valueOf(portalCommonPath.toAbsolutePath()), precision,
+				rtlExcludedPathRegexps, sassCompilerClassName)) {
 
-		String content = new String(Files.readAllBytes(path), charset);
-
-		content = content.replace(s, replacement);
-
-		Files.write(path, content.getBytes(charset));
+			cssBuilder.execute(Collections.singletonList(dirName));
+		}
 	}
 
-	private void _assertMatchesCount(
+	private static void _assertMatchesCount(
 		Pattern pattern, String s, int expectedCount) {
 
 		int count = 0;
@@ -164,87 +151,69 @@ public class CSSBuilderTest {
 		Assert.assertEquals(expectedCount, count);
 	}
 
-	private String _read(String fileName) throws Exception {
-		Path path = Paths.get(fileName);
-
-		String s = new String(Files.readAllBytes(path), StringPool.UTF8);
-
-		return StringUtil.replace(
-			s, StringPool.RETURN_NEW_LINE, StringPool.NEW_LINE);
-	}
-
-	private void _testCssBuilder(String compiler, String portalCommonCssPath)
+	private void _testCSSBuilder(
+			Path portalCommonPath, String sassCompilerClassName)
 		throws Exception {
 
-		try (CSSBuilder cssBuilder = new CSSBuilder(
-				_docrootDirName, false, ".sass-cache/", portalCommonCssPath, 6,
-				new String[0], compiler)) {
+		executeCSSBuilder(
+			"/css", _docrootDirPath, false, ".sass-cache/", portalCommonPath, 6,
+			new String[0], sassCompilerClassName);
 
-			cssBuilder.execute(Arrays.asList(new String[] {"/css"}));
-		}
-
-		String expectedTestContent = _read(
-			_docrootDirName + "/expected/test.css");
-
-		String actualTestContent = _read(
-			_docrootDirName + "/css/.sass-cache/test.css");
+		String expectedTestContent = FileTestUtil.read(
+			_dependenciesDirPath.resolve("expected/test.css"));
+		String actualTestContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test.css"));
 
 		Assert.assertEquals(expectedTestContent, actualTestContent);
 
-		String actualTestCssImportContent = _read(
-			_docrootDirName + "/css/.sass-cache/test_css_import.css");
+		String actualTestImportContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test_css_import.css"));
 
-		_assertMatchesCount(_cssImportPattern, actualTestCssImportContent, 3);
+		_assertMatchesCount(_cssImportPattern, actualTestImportContent, 3);
 
-		String actualTestCssImportRtlContent = _read(
-			_docrootDirName + "/css/.sass-cache/test_css_import_rtl.css");
+		String actualTestImportRtlContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test_css_import_rtl.css"));
 
-		_assertMatchesCount(
-			_cssImportPattern, actualTestCssImportRtlContent, 3);
+		_assertMatchesCount(_cssImportPattern, actualTestImportRtlContent, 3);
 
 		Assert.assertEquals(expectedTestContent, actualTestContent);
 
-		String actualTestPartialContent = _read(
-			_docrootDirName + "/css/.sass-cache/test_partial.css");
+		String actualTestPartialContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test_partial.css"));
 
 		Assert.assertEquals(expectedTestContent, actualTestPartialContent);
 
-		File partialCssFile = new File(
-			Paths.get("/css/.sass-cache/_partial.css").toString());
+		Assert.assertFalse(
+			Files.exists(
+				_docrootDirPath.resolve("css/.sass-cache/_partial.css")));
 
-		Assert.assertFalse(partialCssFile.exists());
-
-		String expectedTestRtlContent = _read(
-			_docrootDirName + "/expected/test_rtl.css");
-
-		String actualTestRtlContent = _read(
-			_docrootDirName + "/css/.sass-cache/test_rtl.css");
+		String expectedTestRtlContent = FileTestUtil.read(
+			_dependenciesDirPath.resolve("expected/test_rtl.css"));
+		String actualTestRtlContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test_rtl.css"));
 
 		Assert.assertEquals(expectedTestRtlContent, actualTestRtlContent);
 
-		String actualTestPartialRtlContent = _read(
-			_docrootDirName + "/css/.sass-cache/test_partial_rtl.css");
+		String actualTestPartialRtlContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test_partial_rtl.css"));
 
 		Assert.assertEquals(
 			expectedTestRtlContent, actualTestPartialRtlContent);
 
-		String expectedUnicodeContent = _read(
-			_docrootDirName + "/expected/test_unicode.css");
-
-		String actualTestUnicodeContent = _read(
-			_docrootDirName + "/css/.sass-cache/test_unicode.css");
+		String expectedUnicodeContent = FileTestUtil.read(
+			_dependenciesDirPath.resolve("expected/test_unicode.css"));
+		String actualTestUnicodeContent = FileTestUtil.read(
+			_docrootDirPath.resolve("css/.sass-cache/test_unicode.css"));
 
 		Assert.assertEquals(expectedUnicodeContent, actualTestUnicodeContent);
 	}
 
-	private static final String _PORTAL_COMMON_CSS_DIR_NAME =
-		"build/portal-common-css";
-
-	private static final String _PORTAL_COMMON_CSS_JAR_FILE_NAME =
-		"build/portal-common-css-jar/com.liferay.frontend.css.common.jar";
-
 	private static final Pattern _cssImportPattern = Pattern.compile(
 		"@import\\s+url\\s*\\(\\s*['\"]?(.+\\.css\\?t=\\d+)['\"]?\\s*\\)\\s*;");
-	private static String _docrootDirName;
+	private static Path _dependenciesDirPath;
+	private static Path _portalCommonDirPath;
+	private static Path _portalCommonJarPath;
+
+	private Path _docrootDirPath;
 
 }